Fix pre-6.0 fake response to checkpoint sync
When sending replica requests for the global checkpoint sync, we skip
sending the request to pre-6.0 nodes since they would not know what to
do with these requests. Yet, we simulate that we received responses from
these nodes. Today, this is done by simulating that they sent us that
their local checkpoint is unassigned sequence number. However, for
pre-6.0 nodes we have introduced a special local checkpoint used in the
global checkpoint tracker for such nodes and that is what we should use
here too. This commit fixes this issue.